ai-ux-review/requirements.txt
fastapi
uvicorn[standard]
pydantic
anthropic
httpx
python-dotenv
pillow
pytest
pytest-asyncio
